refactor pod range proof types to separate submodule
