+++ zk-token-sdk/src/zk_token_elgamal/pod/range_proof.rs
@@ -1,0 +1,31 @@
+use crate::zk_token_elgamal::pod::{Pod, Zeroable};
+
+/// Serialization of range proofs for 64-bit numbers (for `Withdraw` instruction)
+#[derive(Clone, Copy)]
+#[repr(transparent)]
+pub struct RangeProof64(pub [u8; 672]);
+
+// `PodRangeProof64` is a Pod and Zeroable.
+// Add the marker traits manually because `bytemuck` only adds them for some `u8` arrays
+unsafe impl Zeroable for RangeProof64 {}
+unsafe impl Pod for RangeProof64 {}
+
+/// Serialization of range proofs for 128-bit numbers (for `TransferRangeProof` instruction)
+#[derive(Clone, Copy)]
+#[repr(transparent)]
+pub struct RangeProof128(pub [u8; 736]);
+
+// `PodRangeProof128` is a Pod and Zeroable.
+// Add the marker traits manually because `bytemuck` only adds them for some `u8` arrays
+unsafe impl Zeroable for RangeProof128 {}
+unsafe impl Pod for RangeProof128 {}
+
+/// Serialization of range proofs for 128-bit numbers (for `TransferRangeProof` instruction)
+#[derive(Clone, Copy)]
+#[repr(transparent)]
+pub struct RangeProof256(pub [u8; 800]);
+
+// `PodRangeProof256` is a Pod and Zeroable.
+// Add the marker traits manually because `bytemuck` only adds them for some `u8` arrays
+unsafe impl Zeroable for RangeProof256 {}
+unsafe impl Pod for RangeProof256 {}
